SLD-08: type payload families and derive block payload sizes
Replace the closed payload labels with PayloadFamily and emit the existing labels at report and annotation boundaries. Remove decompressed-size duplicates from validated Block and RawBlock; Block::uncomp_sz reads the payload length. Keep the declared size on the unvalidated frame for inflation and admission.

Validation: cargo fmt --all && cargo check -q --workspace --tests passed. Use --no-verify as required by the mission cadence; defer the full gate until all items are complete.

+++ crates/cadmpeg-codec-sldprt/src/writer/tests/sketch_tessellation.rs
@@ -619,7 +619,9 @@ fn semantic_writer_round_trips_all_supported_lanes_together() {
             assert_eq!(block.type_id, 0x77);
         }
         assert!(scan.directory.iter().any(|entry| {
-            entry.name == section && entry.size == block.uncomp_sz && entry.type_id == block.type_id
+            entry.name == section
+                && entry.size as usize == block.uncomp_sz()
+                && entry.type_id == block.type_id
         }));
     }
 }
